refactor: move shared test code to base_test
